Lewd Clicker v0.0.4
Lewd Clicker V0.0.4

New things,

- Added a new start menu.
- Added branding to the new start menu.
- Added nude Nea (which also upgrades your money per click.)

A clicker game with lewd pictures of characters from Dead by Daylight.

The game has multiple, and more being added lewd pictures of characters from the game Dead by Daylight.
If you are a fan of lewd DbD content you should love this game.

Currently has lewd pictures of Kate, Nea, and Meg.

Currently still in development. I'll be working on the game most days adding more and more content.

If you would like to check out older versions or just check out other stuff about the game go to the Github

Note - Most of the art in the game is not mine, If the artiest requests me to remove the art from the game I will be forced to temporarily take the game down.
If you are one of the artiest please contact me on Twitter or email.

To any of the artiest, This game has no intentions of making money at all. I want to try to credit every artiest I can.
 I hope none of you mind me using the art for none profit and just a fun project I wanted to share with others.