Lewd Clicker v0.0.5
Lewd Clicker v0.0.5

New things,
- Added a new way of automaking money. (Duo Partner)
- Added a script that moves the money back and forth depending on how much you have. (If you have $1 the script moves it to the right side, if you have $10 it moves it to the left to fit.)

Changes,
- Changed auto clickers to show the cost of buying one.

Fixes,
- Fixed some button placements to look better.
- Fixed a bug that caused nude Nea's buy button to not disable if you did not have the money.
- A known bug is that the money sometimes falls behind and does not get moved if you spend a good amount of money.

A clicker game with lewd pictures of characters from Dead by Daylight.

The game has multiple, and more being added lewd pictures of characters from the game Dead by Daylight.
If you are a fan of lewd DbD content you should love this game.

Currently has lewd pictures of Kate, Nea, and Meg.

Currently still in development. I'll be working on the game most days adding more and more content.

If you would like to check out older versions or just check out other stuff about the game go to the Github

Note - Most of the art in the game is not mine, If the artiest requests me to remove the art from the game I will be forced to temporarily take the game down.
If you are one of the artiest please contact me on Twitter or email.

To any of the artiest, This game has no intentions of making money at all. I want to try to credit every artiest I can.
 I hope none of you mind me using the art for none profit and just a fun project I wanted to share with others.